requirements.txt
streamlit
openai
numpy
python-dotenv
tiktoken
google-generativeai
